[ATLAS] fix(9dbw): extract shared webhook psycopg fakes — clear Sonar QG
PR #1099 SonarCloud QG failed: new_duplicated_lines_density 4.9%
(threshold 3%). Sole offender — tests/api/webhooks/test_linear_webhook_part1_title.py
at 17.75% (30 duplicated lines): the _FakeCursor/_FakeConn classes were
copy-pasted from test_linear_webhook_kei228.py.

Fix (mirrors the repo's existing tests/scripts/_db_mocks.py pattern):
- New tests/api/webhooks/_webhook_mocks.py — single source of truth for
  FakeCursor + FakeConn.
- test_linear_webhook_part1_title.py imports from it (inline copy removed).
- test_linear_webhook_kei228.py also updated to import from it, so there
  is exactly ONE definition repo-wide — otherwise the duplication would
  just relocate to the new shared module.

No behaviour change: FakeConn keeps the `raise_on_connect` param kei228
declared (unused in both), the `_cur` attribute, and `committed` flag.

Tests: 39 pass (7 part1 + 10 kei228 + 12 kei84 + 10 backfill). ruff
check + format clean.
